old-tests/submission/Microsoft/canvas/canvas_canvasElement_001.htm

<!DOCTYPE HTML> 
<html>
    <head>
        <title>HTML5 Canvas Test: toDataURL() on zero width canvas</title>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com" />
        <link rel="help" href="http://www.w3.org/TR/html5/the-canvas-element.html#the-canvas-element" />
        <meta name="assert" content='If the canvas has a horizontal dimension of zero, then toDataURL() must return the string "data:,".' />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
        <script type="text/javascript">
            setup({explicit_done: true});
            function runTest() 
            {
              test(function()
              {
                var canvas = document.getElementById("canvas1");
                var ctx = canvas.getContext("2d");
                
                // Set the canvas horizontal width to zero.
                canvas.width = 0;

                // toDataURL on zero width canvas.
                var dataURL = canvas.toDataURL();

                // Ensure dataurl represents the empty string in a text/plain resource.
                assert_equals(dataURL, "data:,"); 
              }, 'toDataURL returns "data:"');
              done();
            }
        </script>
    </head>
    <body onload="runTest()">
        <p>Description:  If the canvas has a horizontal dimension of zero, then toDataURL() must return the string "data:,".</p>
        <div id="log"></div>
        <canvas id="canvas1" width="300" height="150">Browser does not support HTML5 Canvas.</canvas>
    </body>
</html>
